+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Fix bug with wrong delivery address in request form. Refs UIREQ-380.
 * Fix menus displaying when duplicating request. Refs UIREQ-385.
 * Fix bug in new request linking back to inventory item record. Fixes UIREQ-349.
+* Fix open request count on item when creating a new request. Fixes UIREQ-343.
 
 ## [1.14.0](https://github.com/folio-org/ui-requests/tree/v1.14.0) (2019-12-05)
 [Full Changelog](https://github.com/folio-org/ui-requests/compare/v1.13.0...v1.14.0)